# swagger.yaml
openapi: 3.0.0
info:
  title: HDR UK API 
  description: API for Tools and artefacts repository.
  version: 0.0.1
servers:
  - url: https://api.www.healthdatagateway.org/api
  - url: http://localhost:3001/api

  - url: https://api.{environment}.healthdatagateway.org:{port}/api
    variables:
      environment:
        default: latest
        description: The Environment name.
      port:
        enum:
          - '443'
        default: '443'
paths:
...
  /v1/tools/:
    get:
      summary: Return List of Tool objects.
      tags:
        - Tools
      parameters:
      - in: query
        name: limit
        required: false
        description: Limit the number of results
        schema:
          type : integer
          example: 3
      - in: query
        name: offset
        required: false
        description: Index to offset the search results
        schema:
          type : integer
          example: 1
      - in: query
        name: q
        required: false
        description: Filter using search query
        schema:
          type : string
          example: epilepsy
      responses: 
        '200':
          description: OK
  /v1/tools/{id}:
    get:
      summary: Returns Tool object.
      tags:
        - Tools
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the tool
          schema:
            type : integer
            format: int64
            minimum: 1
            example: 19009
      responses: 
        '200':
          description: OK
    put:
      summary: Returns edited Tool object.
      security:
        - cookieAuth: []
      tags:
        - Tools
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type : integer
            example: 123
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - name
              properties:
                id:
                  type: number
                type:
                  type: string
                name:
                  type: string
                link:
                  type: string
                description:
                  type: string
                categories:
                  type: object
                  properties:
                    category:
                      type: string
                    programmingLanguage:
                      type: array
                      items:
                        type: string
                    programmingLanguageVersion:
                      type: string
                licence:
                  type: string
                authors:
                  type: array
                  items:
                    type: number
                tags:
                  type: object
                  properties:
                    features:
                      type: array
                      items:
                        type: string
                    topics:
                      type: array
                      items:
                        type: string    
                toolids:
                  type: array
                  items:
                    type: string
              example:   # Sample object
                id: 26542005388306332
                type: 'tool'
                name: 'Research Data TEST EPILEPSY'
                link: 'http://localhost:8080/epilepsy'
                description: 'Epilespy data research description'
                categories: { category: 'API', programmingLanguage: ['Javascript'], programmingLanguageVersion: '1.0.0'}
                licence: 'MIT licence'
                authors: [4495285946631793]
                tags: {features: ['Arbitrage'], topics: ['Epilepsy']}
                toolids: []
      responses: 
        '200':
          description: OK
    patch:
      summary: Change status of Tool object.
      security:
        - cookieAuth: []
      tags:
        - Tools
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the tool
          schema:
            type : integer
            format: int64
            example: 5032687830560181
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - name
              properties:
                id:
                  type: number
                activeflag:
                  type: string
              example:   # Sample object
                id: 662346984100503
                activeflag: 'active'
      responses: 
        '200':
          description: OK
    # Delete requires validation to be implemented first
    # delete:
    #   summary: Delete a tool.
    #   security:
    #     - cookieAuth: []
    #   tags:
    #     - Tools
    #   parameters:
    #     - in: path
    #       name: id
    #       required: true
    #       schema:
    #         type : integer
    #         example: 123
    #   requestBody:
    #     content:
    #       application/json:
    #         schema:      
    #           type: object
    #           required:
    #             - id
    #           properties:
    #             id:
    #               type: number
    #             id: 26542005388306332
    #           example:   # Sample object
    #             id: 26542005388306332
    #   responses: 
    #     '200':
    #       description: OK
    #     '204':
    #       description: No Content
    

  

  
  

  # /v1/tools/get:
  #   get:
  #     summary: Tool search.
  #     tags:
  #       - Tool
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/tools/get/admin:
  #   get:
  #     summary: Tool search admins.
  #     tags:
  #       - Tool
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/tools/status:
  #   put:
  #     summary: Tool update status.
  #     tags:
  #       - Tool
  #     responses: 
  #       '200':
  #         description: OK

    # /v1/project/get:
  #   get:
  #     summary: Project search.
  #     tags:
  #       - Project
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/project/get/admin:
  #   get:
  #     summary: Project search admins.
  #     tags:
  #       - Project
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/project/status:
  #   put:
  #     summary: Project update status.
  #     tags:
  #       - Project
  #     responses: 
  #       '200':
  #         description: OK

     # /v1/paper/get:
  #   get:
  #     summary: Paper search.
  #     tags:
  #       - Paper
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/paper/get/admin:
  #   get:
  #     summary: Paper search admins.
  #     tags:
  #       - Paper
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/paper/status:
  #   put:
  #     summary: Paper update status.
  #     tags:
  #       - Paper
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/users/{userId}:
  #   get:
  #     summary: Find user by id.
  #     tags:
  #       - Users
  #   parameters:
  #     - name: userId
  #       in: path
  #       required: true
  #       description: The ID of the user
  #       schema:
  #         type : string
  #         example: b2475d19-75fe-422b-9d0c-c53cdfff8230
  #   responses:
  #     '200':
  #       description: OK
  
  # /v1/users:
  #   get:
  #     summary: Find all users
  #     tags:
  #       - users
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/auth/register:
  #   post:
  #     summary: Register User.
  #     tags:
  #       - Users
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/auth/login:
  #   post:
  #     summary: Authenticate User.
  #     tags:
  #       - Auth   
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/auth/logout:
  #   get:
  #     summary: Logout.
  #     tags:
  #       - Auth
  #     responses: 
  #       '200':
  #         description: OK
  
  # /v1/auth/status:
  #   get:
  #     summary: Auth status.
  #     tags:
  #       - Auth
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/tools/add:
  #   post:
  #     summary: Add a tool.
  #     tags:
  #       - Tool
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/tools/edit:
  #   put:
  #     summary: Edit a tool.
  #     tags:
  #       - Tool
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/tools/delete:
  #   delete:
  #     summary: delete a tool.
  #     tags:
  #       - Tool
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/project/add:
  #   post:
  #     summary: Add a project.
  #     tags:
  #       - Project
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/project/edit:
  #   put:
  #     summary: Edit a project.
  #     tags:
  #       - Project
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/project/delete:
  #   delete:
  #     summary: delete a project.
  #     tags:
  #       - Project
  #     responses: 
  #       '200':
  #         description: OK

   # /v1/paper/add:
  #   post:
  #     summary: Add a paper.
  #     tags:
  #       - Paper
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/paper/edit:
  #   put:
  #     summary: Edit a paper.
  #     tags:
  #       - Paper
  #     responses: 
  #       '200':
  #         description: OK

  # /v1/paper/delete:
  #   delete:
  #     summary: delete a paper.
  #     tags:
  #       - Paper
  #     responses: 
  #       '200':
  #         description: OK
  
  # /v1/messages/{personId}:
  #   get:
  #     summary: Find messages by person.
  #     tags:
  #       - Messages
  #     parameters:
  #       - name: personId
  #         in: path
  #         required: true
  #         description: The ID of the person
  #         schema:
  #           type : string
  #           example: b2475d19-75fe-422b-9d0c-c53cdfff8230

  # /v1/messages/admin/{personId}:
  #   get:
  #     summary: find messages for admin by person id
  #     tags:
  #       - Messages
  #     parameters:
  #       - name: personId
  #         in: path
  #         required: true
  #         description: The ID of the person
  #         schema:
  #           type : string
  #           example: b2475d19-75fe-422b-9d0c-c53cdfff8230

  # /v1/reviews:
  #   get:
  #     summary: find all reviews.
  #     tags:
  #       - Reviews

  # /v1/reviews/pending:
  #   get:
  #     summary: find all pending reviews.
  #     tags:
  #       - Reviews

  # /v1/reviews/admin/pending:
  #   get:
  #     summary: find all pending reviews for admins.
  #     tags:
  #       - Reviews

  # /v1/discourse/topic/tool/{toolId}:
  #   put:
  #     summary: This route creates a new topic in Discourse if the tool exists and is active. Returns an object with the link to the new Discourse topic.
  #     tags:
  #       - Discourse
  #     parameters:
  #       - name: toolId
  #         in: path
  #         required: true
  #         description: Tool id